Update account registrations list.
