+++ mhr_api/src/mhr_api/models/db2/manuhome.py
@@ -16,7 +16,6 @@ from enum import Enum
 from http import HTTPStatus
 
 from flask import current_app
-from sqlalchemy.sql import text
 
 from mhr_api.exceptions import BusinessException, DatabaseException, ResourceErrorCodes
 from mhr_api.models import utils as model_utils
@@ -30,50 +29,11 @@ from .owner import Db2Owner
 from .owngroup import Db2Owngroup
 
 
-QUERY_ACCOUNT_ADD_REGISTRATION = """
-SELECT mh.mhregnum, mh.mhstatus, d.regidate, TRIM(d.name), TRIM(d.olbcfoli), TRIM(d.docutype),
-       (SELECT XMLSERIALIZE(XMLAGG ( XMLELEMENT ( NAME "owner", TRIM(o2.ownrname))) AS CLOB)
-          FROM owner o2, owngroup og2
-         WHERE o2.manhomid = mh.manhomid
-           AND og2.manhomid = mh.manhomid
-           AND og2.owngrpid = o2.owngrpid
-           AND og2.status IN ('3', '4')) as owner_names,
-       TRIM(d.bcolacct)
-  FROM manuhome mh, document d
- WHERE mh.mhregnum = :query_mhr_number
-   AND mh.mhregnum = d.mhregnum
-   AND mh.regdocid = d.documtid
-"""
-QUERY_ACCOUNT_REGISTRATIONS = """
-SELECT mh.mhregnum, mh.mhstatus, d.regidate, TRIM(d.name), TRIM(d.olbcfoli), TRIM(d.docutype),
-       (SELECT XMLSERIALIZE(XMLAGG ( XMLELEMENT ( NAME "owner", TRIM(o2.ownrname))) AS CLOB)
-          FROM owner o2, owngroup og2
-         WHERE o2.manhomid = mh.manhomid
-           AND og2.manhomid = mh.manhomid
-           AND og2.owngrpid = o2.owngrpid
-           AND og2.status IN ('3', '4')) as owner_names,
-       TRIM(d.bcolacct)
-  FROM manuhome mh, document d
- WHERE mh.mhregnum IN (?)
-   AND mh.mhregnum = d.mhregnum
-   AND mh.regdocid = d.documtid
-ORDER BY d.regidate DESC
-"""
-DOC_ID_COUNT_QUERY = """
-SELECT COUNT(documtid)
-  FROM document
- WHERE documtid = :query_value
-"""
-REGISTRATION_DESC_NEW = 'Manufactured Home Registration'
 LEGACY_STATUS_DESCRIPTION = {
     'R': 'ACTIVE',
     'E': 'EXEMPT',
     'D': 'DRAFT',
     'C': 'HISTORICAL'
-}
-LEGACY_REGISTRATION_DESCRIPTION = {
-    '101': REGISTRATION_DESC_NEW,
-    'CONV': REGISTRATION_DESC_NEW
 }
 DOCUMENT_TYPE_REG = '101'
 
@@ -308,94 +268,6 @@ class Db2Manuhome(db.Model):
         return man_home
 
     @classmethod
-    def find_summary_by_mhr_number(cls, mhr_number: str):
-        """Return the MH registration summary information matching the MHR number."""
-        summary = None
-        try:
-            query = text(QUERY_ACCOUNT_ADD_REGISTRATION)
-            result = db.get_engine(current_app, 'db2').execute(query, {'query_mhr_number': mhr_number})
-            rows = result.fetchall()
-            if rows is not None:
-                for row in rows:
-                    summary = Db2Manuhome.__build_summary(row, True)
-        except Exception as db_exception:   # noqa: B902; return nicer error
-            current_app.logger.error('DB2 find_summary_by_mhr_number exception: ' + str(db_exception))
-            raise DatabaseException(db_exception)
-        return summary
-
-    @classmethod
-    def find_summary_by_account_mhr_numbers(cls, mhr_list):
-        """Return a list of MH registration summary information matching the account MHR numbers."""
-        summary_list = []
-        if not mhr_list:
-            return summary_list
-        try:
-            mhr_numbers: str = ''
-            count = 0
-            for mhr in mhr_list:
-                mhr_number = mhr['mhr_number']
-                count += 1
-                if count > 1:
-                    mhr_numbers += ','
-                mhr_numbers += f"'{mhr_number}'"
-            query_text = QUERY_ACCOUNT_REGISTRATIONS.replace('?', mhr_numbers)
-            # current_app.logger.debug('Executing query=' + query_text)
-            query = text(query_text)
-            result = db.get_engine(current_app, 'db2').execute(query)
-            rows = result.fetchall()
-            if rows is not None:
-                for row in rows:
-                    summary_list.append(Db2Manuhome.__build_summary(row, False))
-        except Exception as db_exception:   # noqa: B902; return nicer error
-            current_app.logger.error('DB2 find_summary_by_mhr_number exception: ' + str(db_exception))
-            raise DatabaseException(db_exception)
-        return summary_list
-
-    @classmethod
-    def get_doc_id_count(cls, doc_id):
-        """Execute a DB2 query to count existing document id (must not exist check)."""
-        try:
-            query = text(DOC_ID_COUNT_QUERY)
-            result = db.get_engine(current_app, 'db2').execute(query, {'query_value': doc_id})
-            row = result.first()
-            exist_count = int(row[0])
-            current_app.logger.debug(f'Existing doc id count={exist_count}.')
-            return exist_count
-        except Exception as db_exception:   # noqa: B902; return nicer error
-            current_app.logger.error('DB2 get_doc_id_count exception: ' + str(db_exception))
-            raise DatabaseException(db_exception)
-
-    @classmethod
-    def __build_summary(cls, row, add_in_user_list: bool = True):
-        """Build registration summary from query result."""
-        mhr_number = str(row[0])
-        timestamp = row[2]
-        owners = str(row[6])
-        owners = owners.replace('<owner>', '')
-        owners = owners.replace('</owner>', '\n')
-        owner_list = owners.split('\n')
-        owner_names = ''
-        for name in owner_list:
-            if name.strip() != '':
-                owner_names += Db2Manuhome.get_summary_name(name) + '\n'
-        summary = {
-            'mhrNumber': mhr_number,
-            'registrationDescription': LEGACY_REGISTRATION_DESCRIPTION.get(str(row[5]),
-                                                                           REGISTRATION_DESC_NEW),
-            'username': str(row[7]),
-            'statusType': LEGACY_STATUS_DESCRIPTION.get(str(row[1])),
-            'createDateTime': model_utils.format_local_ts(timestamp),
-            'submittingParty': str(row[3]),
-            'clientReferenceId': str(row[4]),
-            'ownerNames': owner_names,
-            'path': ''
-        }
-        if add_in_user_list:
-            summary['inUserList'] = False
-
-        return summary
-
-    @classmethod
     def __sort_notes(cls, notes):
         """Sort notes by registration timesamp."""
         notes.sort(key=Db2Manuhome.__sort_key_notes_ts, reverse=True)
@@ -471,16 +343,3 @@ class Db2Manuhome(db.Model):
                 group = Db2Owngroup.create_from_registration(registration, group, (i + 1))
                 manuhome.reg_owner_groups.append(group)
         return manuhome
-
-    @staticmethod
-    def get_summary_name(db2_name: str):
-        """Get an individual name json from a DB2 legacy name."""
-        last = db2_name[0:24].strip()
-        first = db2_name[25:].strip()
-        middle = None
-        if len(db2_name) > 40:
-            first = db2_name[25:38].strip()
-            middle = db2_name[39:].strip()
-        if middle:
-            return first + ' ' + middle + ' ' + last
-        return first + ' ' + last
